requirements.txt
torch 
torchvision
numpy
matplotlib
Path
PIL
PyYAML